Retrieve survey list from backstage api
Added controllers folder with a function to return the survey
list from the survey service. Also added relevant exceptions/
error handlers for this process
