+++ response_operations_ui/__init__.py
@@ -24,5 +24,6 @@ app.url_map.strict_slashes = False
 logger_initial_config(service_name='response-operations-ui', log_level=app.config['LOGGING_LEVEL'])
 
 
+from response_operations_ui import error_handlers  # NOQA # pylint: disable=wrong-import-position
 from response_operations_ui.views import surveys  # NOQA # pylint: disable=wrong-import-position
 from response_operations_ui.views import info  # NOQA # pylint: disable=wrong-import-position
