+++ response_operations_ui/error_handlers.py
@@ -1,0 +1,23 @@
+import logging
+
+from requests import ConnectionError
+from structlog import wrap_logger
+
+from response_operations_ui import app
+from response_operations_ui.exceptions.exceptions import ApiError
+
+
+logger = wrap_logger(logging.getLogger(__name__))
+
+
+@app.errorhandler(ApiError)
+def api_error(error):
+    logger.error('Api failed to retrieve required data',
+                 url=error.url, status_code=str(error.status_code))
+    return "FAIL"
+
+
+@app.errorhandler(ConnectionError)
+def connection_error(error):
+    logger.error('Failed to connect to external service', url=error.request.url)
+    return "FAIL"
