+++ response_operations_ui/exceptions/exceptions.py
@@ -1,0 +1,5 @@
+class ApiError(Exception):
+
+    def __init__(self, response):
+        self.url = response.url
+        self.status_code = response.status_code
