+++ response_operations_ui/templates/layouts/base.html
@@ -15,7 +15,7 @@
     <!--{% assets "js_all" %}-->
     <!--<script type="text/javascript" src="{{ ASSET_URL }}"></script>-->
     <!--{% endassets %}-->
-    <!--<link href="/images/favicon.ico" rel="Shortcut Icon" />-->
+    <link href="/static/images/favicon.ico" rel="Shortcut Icon" />
   </head>
 
   <body>
