+++ response_operations_ui/views/surveys.py
@@ -1,11 +1,10 @@
-import json
 import logging
 
 from flask import render_template
-import requests
 from structlog import wrap_logger
 
 from response_operations_ui import app
+from response_operations_ui.controllers import survey_controllers
 
 
 logger = wrap_logger(logging.getLogger(__name__))
@@ -13,17 +12,15 @@ logger = wrap_logger(logging.getLogger(__name__))
 
 @app.route('/', methods=['GET'])
 def view_surveys():
-    return render_template('surveys.html')
 
+    # survey_list = [
+    #     {
+    #         "id": "cb0711c3-0ac8-41d3-ae0e-567e5ea1ef87",
+    #         "shortName": "BRES",
+    #         "longName": "Business Register and Employment Survey",
+    #         "surveyRef": "221"
+    #     }
+    # ]
 
-def get_surveys_list():
-    logger.debug('Retrieving surveys list')
-    url = '{}/{}'.format(app.config['BACKSTAGE_API_URL'], 'surveys')
-
-    response = requests.get(url)
-    if response.status_code != 200:
-        logger.error('Failed to retrieve surveys list')
-        raise Exception
-
-    logger.debug('Successfully retrieved surveys list')
-    return json.loads(response.text)
+    survey_list = survey_controllers.get_surveys_list()
+    return render_template('surveys.html', survey_list=survey_list)
